[ROCm] Repair the HIP build for current ROCm wave-size macros
LC already ships an author-maintained HIP path, but it gated everything
on __AMDGCN_WAVEFRONT_SIZE, a predefined macro that ROCm 7.x no longer
provides. With the macro gone the AMD shim block in include/macros.h
never compiled (so __trap, atomic*_block, and the cuda::std alias were
undeclared) and the WS wave-width define silently fell back to 32 on
wave64 hardware, which would corrupt the warp-collective zero- and
repetition-elimination paths. This restores the HIP build on a current
ROCm without changing the CUDA build.

Review include/macros.h first: the shim block is re-gated on
__HIP_PLATFORM_AMD__ (defined in both host and device passes), <limits>
is pulled in so the cuda::std::numeric_limits alias resolves, and the
__syncwarp shim is now conditional because recent ROCm provides
__syncwarp natively. Then the wave-width selector in the four framework
sources and the two Python generators that emit include/consts.h: wave64
is selected on GCN/CDNA via the device-pass __GFX8__/__GFX9__ macros and
wave32 elsewhere, which keeps the serialized bitstream format identical
across wave widths (verified by decoding gfx90a output on the CPU
reference). framework.h casts warpSize to int for device printf, and
d_QUANT_INOA_0_f64.h gains the thrust includes its f32 sibling already
has (a latent omission that compiled on CUDA only by transitive luck).

The CUDA path is unchanged: none of __HIP_PLATFORM_AMD__/__GFX8__/
__GFX9__ are defined for nvcc, so WS stays 32 and the shim block stays
disabled exactly as before.

This work was authored with the assistance of Claude, an AI assistant.

Test Plan:
Generated, hipified, and built the framework for gfx90a on ROCm 7.2.1:

```
./generate_Device_LC-Framework.py
hipify-perl -inplace lc.cu lc.h
for h in framework.h $(find include components preprocessors verifiers -name '*.h'); do hipify-perl -inplace "$h"; done
hipcc -O3 --offload-arch=gfx90a -ffp-contract=off -DUSE_GPU -I. -std=c++17 -o lc lc.cu
```

Confirmed the device pass selects WS=64 on gfx90a (device asm stores 64).
Ran the built-in lossless round-trip (AL mode) across the warp-collective
pipelines (RZE_4, RZE_1, RLE_4, BIT_4 RLE_4, RRE_4 RZE_4): all pass
bit-for-bit. Ran the TS self-test over all component pairs with no
failures. Lossy quantizers verified within bound (QUANT_ABS, QUANT_NOA,
QUANT_INOA_0_f64 with MAXABS/MAXNOA verifiers), exercising the thrust and
cuda::std paths. Cross-device format gate: data compressed on gfx90a
(wave64) decompressed correctly on the CPU reference build (wave32),
byte-for-byte identical to the original.

+++ framework.h
@@ -218,7 +218,7 @@ static void h_preprocess_decode(long long& hpredecsize, byte*& hpredecdata, std:
 #ifdef USE_GPU
 static __global__ void initBestSize(unsigned short* const bestSize, const int chunks)
 {
-  if ((threadIdx.x == 0) && (WS != warpSize)) {printf("ERROR: WS must be %d\n\n", warpSize); __trap();}  // debugging only
+  if ((threadIdx.x == 0) && (WS != warpSize)) {printf("ERROR: WS must be %d\n\n", (int)warpSize); __trap();}  // debugging only
   for (int i = threadIdx.x; i < chunks; i += TPB) {
     bestSize[i] = CS;
   }
